fix(persistence): use i32 for Postgres int4 manifest counters
The bulk_manifests counter columns (total_entries, processed_entries,
failed_entries) are declared INTEGER (int4), but the Postgres backend
read them as i64 in get_manifest and bound i64 params in the counts
UPDATE. tokio-postgres rejects the int4<->i64 width conversion and
panicked the request handler, dropping the connection mid-response.
This surfaced as `curl: (52) Empty reply from server` when polling
$bulk-submit-status, failing the external smoke workflow. SQLite was
unaffected because it is dynamically typed.

Read and bind these three columns as i32 to match the schema.

Tests: cargo test -p helios-persistence --lib (655 passed); fmt and
CI-style clippy clean.

+++ crates/persistence/src/backends/postgres/bulk_submit.rs
@@ -540,9 +540,9 @@ impl BulkSubmitProvider for PostgresBackend {
         let replaces_manifest_url: Option<String> = row.get(1);
         let status_str: String = row.get(2);
         let added_at: chrono::DateTime<Utc> = row.get(3);
-        let total: i64 = row.get(4);
-        let processed: i64 = row.get(5);
-        let failed: i64 = row.get(6);
+        let total: i32 = row.get(4);
+        let processed: i32 = row.get(5);
+        let failed: i32 = row.get(6);
 
         let status: ManifestStatus = status_str
             .parse()
@@ -701,9 +701,9 @@ impl BulkSubmitProvider for PostgresBackend {
                     failed_entries = failed_entries + $3
                  WHERE tenant_id = $4 AND submitter = $5 AND submission_id = $6 AND manifest_id = $7",
                 &[
-                    &(results.len() as i64),
-                    &(results.iter().filter(|r| r.is_success()).count() as i64),
-                    &(error_count as i64),
+                    &(results.len() as i32),
+                    &(results.iter().filter(|r| r.is_success()).count() as i32),
+                    &(error_count as i32),
                     &tenant_id,
                     &submission_id.submitter.as_str(),
                     &submission_id.submission_id.as_str(),
